refactor(pip_requirements): Move flags extraction to common.ts
The pip-compile manager will need to extract flags from pip requirements
files in #28959, so move that code into a common.ts
